requirements.txt
Pillow==9.1.1
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.11.0+cpu
torchvision==0.12.0
gunicorn ==20.1.0
fastapi==0.79.0
uvicorn==0.18.2
python-multipart
pickle-mixin